update: create selector component
